Move blockAtlas loading to separate function
Incase the textures ever want to be changed in-game (in the case of texture packs or otherwise), then loadBlocksAtlas() will read resources/blocks.png from file again. In addition it will set its filter to nearest neighbor so the blocks' pixels stay crisp when zoomed, and get the number of blocks in the image (if the blocks are 16px).